fix: unsigned dtype handling in TUI, filter predicates, and Calcite (#216)
Follow-up to #208 (CSV export): the remaining consumers still read unsigned
integer columns through raw signed getters, so high-half values (bit set at the
width's sign position) were mishandled. Per site:

- cli tui GridRender / InspectorRender: U64/U32 now render via
  Long/Integer.toUnsignedString; U16/U8 render via the dtype-aware getInt
  (zero-extend). Display bug only, but visible on any unsigned column.
- cli FilterCommand.compareValue: the worst class — silent wrong query results.
  Filter predicates now widen U8/U16/U32 losslessly into a non-negative long and
  compare U64 with Long.compareUnsigned, so `magnesium >= 130` matches its
  high-half U8 rows instead of dropping them. Made package-private + unit-tested.
- calcite VortexTable: unsigned columns now map to the next wider signed SQL type
  so the declared type holds their full range (U8->SMALLINT, U16->INTEGER,
  U32->BIGINT) and value() widens accordingly. U64 has no wider SQL integer, so
  it stays BIGINT and value() throws VortexException for values >= 2^63 rather
  than surfacing a two's-complement negative.
- calcite VortexAggregates.scanSum: U32 elements widen via toUnsignedLong;
  U64 elements >= 2^63 throw (aligned with the VortexTable decision). Branch-split
  on the loop-invariant unsigned flag keeps the signed body vectorizable.

U64 decision: mapped to signed BIGINT because Calcite has no wider SQL integer;
a documented VortexException for the unrepresentable high half is chosen over a
silent negative (loud beats wrong) — both when a row is read and when it is
summed.

Closes #216

+++ calcite/src/test/java/io/github/dfa1/vortex/calcite/VortexAdapterCoverageTest.java
@@ -82,11 +82,13 @@ class VortexAdapterCoverageTest {
         // Given / When
         RelDataType rowType = new VortexTable(file).getRowType(new JavaTypeFactoryImpl());
 
-        // Then — one SqlTypeName per logical type
+        // Then — one SqlTypeName per logical type. Unsigned integers map to the next wider signed
+        // SQL type so the declared type holds their full range (#216); U64 is the exception (no
+        // wider SQL integer) and stays BIGINT with a value-side overflow guard.
         Map<String, SqlTypeName> expected = Map.ofEntries(
-                Map.entry("i8", SqlTypeName.TINYINT), Map.entry("u8", SqlTypeName.TINYINT),
-                Map.entry("i16", SqlTypeName.SMALLINT), Map.entry("u16", SqlTypeName.SMALLINT),
-                Map.entry("i32", SqlTypeName.INTEGER), Map.entry("u32", SqlTypeName.INTEGER),
+                Map.entry("i8", SqlTypeName.TINYINT), Map.entry("u8", SqlTypeName.SMALLINT),
+                Map.entry("i16", SqlTypeName.SMALLINT), Map.entry("u16", SqlTypeName.INTEGER),
+                Map.entry("i32", SqlTypeName.INTEGER), Map.entry("u32", SqlTypeName.BIGINT),
                 Map.entry("i64", SqlTypeName.BIGINT), Map.entry("u64", SqlTypeName.BIGINT),
                 Map.entry("f32", SqlTypeName.REAL), Map.entry("f64", SqlTypeName.DOUBLE),
                 Map.entry("s", SqlTypeName.VARCHAR), Map.entry("b", SqlTypeName.BOOLEAN));
